Removed redundant duplicate count of failing dies
Previously, if a die is found in more than one class (besides the first class folder), then on the failing count number at the wafer map would count these dies multiple times. Now it will check if die has already been counted for, and if so, to continue to next die name.